refactor: unify glucose measurement interface
