+++ components/GlucoseChart.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text, StyleSheet, Dimensions } from 'react-native';
 import { LineChart } from 'react-native-chart-kit';
-import { GlucoseMeasurement } from '@/utils/storage';
+import { GlucoseMeasurement } from '@/types/glucose';
 import { getGlucoseStatus } from '@/utils/glucose';
 
 interface GlucoseChartProps {
